requirements.txt
sqlalchemy==2.0.13
requests==2.30.0
pydantic==1.10.7
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
apscheduler==3.10.1



